sort/sort.go を修正しました。
引数の slice の型を int から float64 に変更しました。

+++ sort/sort.go
@@ -5,14 +5,14 @@ import (
 	"sort"
 )
 
-func Sort(s1 []int) []int{
+func Sort(s1 []float64) []float64{
 
 	// 昇順
-	sort.Sort(sort.IntSlice(s1))
+	sort.Sort(sort.Float64Slice(s1))
 	fmt.Printf("len:%d cap:%d v:%v\n", len(s1), cap(s1), s1)
 
 	// 降順
-	sort.Sort(sort.Reverse(sort.IntSlice(s1)))
+	sort.Sort(sort.Reverse(sort.Float64Slice(s1)))
 	fmt.Printf("len:%d cap:%d v:%v\n", len(s1), cap(s1), s1)
 
 	return s1
